pyproject.toml
[project]
name = "aiimage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.49.0",
    "pillow>=12.0.0",
    "python-dotenv>=1.2.1",
]
